Added support for more vertex list primitives such as triangle strips, line strips, and quad strips. Currently not implemented for all renderers except for the previously existing triangles and triangle fan. This feature can be accessed via the new driver method drawVertexPrimitiveList using vertex and index pointers and a specification of the vertex and primitive types. Even simpler access is provided through mesh buffers which also hold a primitives type now and are automatically drawn with the correct method. All the other drawTriangle* methods are now inlined convenience methods provided by the NullDriver. Changed SkyBox to render with triangle fans which requires less indices. Another fix for the OpenGL specular colors for which an old bug showed up again that had been invisible due to black specular highlights in the last release. The Linux video mode list now contains depth information, though only an educated guess instead of proper values. The software driver clipping has been changed, resulting in almost complete rendering of the meshes, though probably a little slower. Moreover the texture coordinates seem to be messed up and drift upon camera motion.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@151 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CVideoModeList.cpp
@@ -82,7 +82,7 @@ void CVideoModeList::addMode(const core::dimension2d<s32>& size, s32 depth)
 	m.depth = depth;
 	m.size = size;
 
-	for (s32 i=0; i<(s32)VideoModes.size(); ++i)
+	for (u32 i=0; i<VideoModes.size(); ++i)
 	{
 		if (VideoModes[i] == m)
 			return;
